AI Module 1/Module 1/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
requests>=2.28.0
python-dotenv>=1.0.0
